2016.08.22 13:48
- 버블소트 완성(필요할 시, 이미지로 대체)
- 문제버그 해결
