+++ Season.lua
@@ -18,9 +18,9 @@ function CheckSeason()
     elseif stageLevel == 2 then
       CreateSummer()
     elseif stageLevel == 3 then
-      --CreateFall()
-      stageLevel = 7
-      CreateBossCastle()
+      CreateFall()
+      --stageLevel = 7
+      --CreateBossCastle()
     elseif stageLevel == 4 then
       CreateWinter()
   end
